Update Ansible and npm
This project does not list any npm "dependencies"--only
"devDependencies". The Ansible "module" for npm uses `npm list` to
determine when module installation is necessary.

In prior releases of Ansible, this was achieved via the command `npm
list --json`. This was apparently not designed for projects which omit
"dependencies" because this project triggered an inappropriate code
path. npm package installation proceeded as expected, however, so the
bug was not apparent.

A recent change to the internals of the Ansible module introduced the
`--long` flag to `npm list` [1]. This invalidated accidental behavior
which previously triggered module installation, preventing any modules
from being installed.

A separate bug in the npm Ansible module interferes with module
installation for projects which list some "devDependencies" but zero
"dependencies" [2].

Update the version of Node.js (and therefore npm) to the latest version
as this is known to integrate with the latest release of Ansible. Update
the project documentation to reflect the new supported versions. Declare
project dependencies as npm "dependencies" (i.e. not "devDependencies")
in order to work around a known issue and to more accurately describe the
context in which modules must be installed.

[1] https://github.com/ansible/ansible/commit/cea681a
[2] https://github.com/ansible/ansible/issues/44346